Refactor database client remote function structure
