+++ stdlib/database/sql/src/main/java/org/ballerinalang/database/sql/SQLDatasourceUtils.java
@@ -21,44 +21,33 @@ import org.ballerinalang.bre.Context;
 import org.ballerinalang.bre.bvm.BLangVMErrors;
 import org.ballerinalang.connector.api.BLangConnectorSPIUtil;
 import org.ballerinalang.model.ColumnDefinition;
-import org.ballerinalang.model.types.BArrayType;
-import org.ballerinalang.model.types.BField;
-import org.ballerinalang.model.types.BStructureType;
-import org.ballerinalang.model.types.BType;
 import org.ballerinalang.model.types.BTypes;
 import org.ballerinalang.model.types.TypeKind;
 import org.ballerinalang.model.types.TypeTags;
-import org.ballerinalang.model.values.BBoolean;
-import org.ballerinalang.model.values.BDecimal;
 import org.ballerinalang.model.values.BError;
-import org.ballerinalang.model.values.BFloat;
 import org.ballerinalang.model.values.BInteger;
 import org.ballerinalang.model.values.BMap;
 import org.ballerinalang.model.values.BRefType;
-import org.ballerinalang.model.values.BString;
 import org.ballerinalang.model.values.BValue;
-import org.ballerinalang.model.values.BValueArray;
 import org.ballerinalang.util.exceptions.BallerinaException;
+import org.ballerinalang.util.observability.ObservabilityConstants;
+import org.ballerinalang.util.observability.ObserveUtils;
+import org.ballerinalang.util.observability.ObserverContext;
 import org.ballerinalang.util.transactions.BallerinaTransactionContext;
 import org.ballerinalang.util.transactions.TransactionLocalContext;
 import org.ballerinalang.util.transactions.TransactionResourceManager;
 import org.ballerinalang.util.transactions.TransactionUtils;
 
 import java.io.BufferedReader;
-import java.io.ByteArrayInputStream;
 import java.io.File;
 import java.io.IOException;
 import java.io.Reader;
-import java.io.StringReader;
-import java.math.BigDecimal;
 import java.nio.charset.Charset;
 import java.sql.Array;
 import java.sql.Blob;
-import java.sql.CallableStatement;
 import java.sql.Clob;
 import java.sql.Connection;
 import java.sql.Date;
-import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.ResultSetMetaData;
 import java.sql.SQLException;
@@ -68,15 +57,12 @@ import java.sql.Time;
 import java.sql.Timestamp;
 import java.sql.Types;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Base64;
 import java.util.Calendar;
-import java.util.GregorianCalendar;
 import java.util.HashSet;
-import java.util.Iterator;
 import java.util.List;
 import java.util.Locale;
-import java.util.Map;
+import java.util.Optional;
 import java.util.Set;
 import java.util.StringJoiner;
 import java.util.TimeZone;
@@ -93,774 +79,15 @@ import javax.transaction.xa.XAResource;
  */
 public class SQLDatasourceUtils {
 
-    private static final String POSTGRES_DOUBLE = "float8";
     public static final String POSTGRES_OID_COLUMN_TYPE_NAME = "oid";
-    private static final int ORACLE_CURSOR_TYPE = -10;
     private static final String POOL_MAP_KEY = UUID.randomUUID().toString();
 
-    public static void setIntValue(PreparedStatement stmt, BValue value, int index, int direction, int sqlType) {
-        Integer val = obtainIntegerValue(value);
-        try {
-            if (Constants.QueryParamDirection.IN == direction) {
-                if (val == null) {
-                    stmt.setNull(index + 1, sqlType);
-                } else {
-                    stmt.setInt(index + 1, val);
-                }
-            } else if (Constants.QueryParamDirection.INOUT == direction) {
-                if (val == null) {
-                    stmt.setNull(index + 1, sqlType);
-                } else {
-                    stmt.setInt(index + 1, val);
-                }
-                ((CallableStatement) stmt).registerOutParameter(index + 1, sqlType);
-            } else if (Constants.QueryParamDirection.OUT == direction) {
-                ((CallableStatement) stmt).registerOutParameter(index + 1, sqlType);
-            } else {
-                throw new BallerinaException("invalid direction for the parameter with index: " + index);
-            }
-        } catch (SQLException e) {
-            throw new BallerinaException("error in set integer to statement: " + e.getMessage(), e);
-        }
-    }
-
-    public static void setSmallIntValue(PreparedStatement stmt, BValue value, int index, int direction, int sqlType) {
-        Integer val = obtainIntegerValue(value);
-        try {
-            if (Constants.QueryParamDirection.IN == direction) {
-                if (val == null) {
-                    stmt.setNull(index + 1, sqlType);
-                } else {
-                    stmt.setShort(index + 1, val.shortValue());
-                }
-            } else if (Constants.QueryParamDirection.INOUT == direction) {
-                if (val == null) {
-                    stmt.setNull(index + 1, sqlType);
-                } else {
-                    stmt.setShort(index + 1, val.shortValue());
-                }
-                ((CallableStatement) stmt).registerOutParameter(index + 1, sqlType);
-            } else if (Constants.QueryParamDirection.OUT == direction) {
-                ((CallableStatement) stmt).registerOutParameter(index + 1, sqlType);
-            } else {
-                throw new BallerinaException("invalid direction for the parameter with index: " + index);
-            }
-        } catch (SQLException e) {
-            throw new BallerinaException("error in set integer to statement: " + e.getMessage(), e);
-        }
-    }
-
-    public static void setStringValue(PreparedStatement stmt, BValue value, int index, int direction, int sqlType) {
-        try {
-            if (Constants.QueryParamDirection.IN == direction) {
-                if (value == null) {
-                    stmt.setNull(index + 1, sqlType);
-                } else {
-                    stmt.setString(index + 1, value.stringValue());
-                }
-            } else if (Constants.QueryParamDirection.INOUT == direction) {
-                if (value == null) {
-                    stmt.setNull(index + 1, sqlType);
-                } else {
-                    stmt.setString(index + 1, value.stringValue());
-                }
-                ((CallableStatement) stmt).registerOutParameter(index + 1, sqlType);
-            } else if (Constants.QueryParamDirection.OUT == direction) {
-                ((CallableStatement) stmt).registerOutParameter(index + 1, sqlType);
-            } else {
-                throw new BallerinaException("invalid direction for the parameter with index: " + index);
-            }
-        } catch (SQLException e) {
-            throw new BallerinaException("error in set string to statement: " + e.getMessage(), e);
-        }
-    }
-
-    public static void setNStringValue(PreparedStatement stmt, BValue value, int index, int direction, int sqlType) {
-        try {
-            if (Constants.QueryParamDirection.IN == direction) {
-                if (value == null) {
-                    stmt.setNull(index + 1, sqlType);
-                } else {
-                    stmt.setNString(index + 1, value.stringValue());
-                }
-            } else if (Constants.QueryParamDirection.INOUT == direction) {
-                if (value == null) {
-                    stmt.setNull(index + 1, sqlType);
-                } else {
-                    stmt.setNString(index + 1, value.stringValue());
-                }
-                ((CallableStatement) stmt).registerOutParameter(index + 1, sqlType);
-            } else if (Constants.QueryParamDirection.OUT == direction) {
-                ((CallableStatement) stmt).registerOutParameter(index + 1, sqlType);
-            } else {
-                throw new BallerinaException("invalid direction for the parameter with index: " + index);
-            }
-        } catch (SQLException e) {
-            throw new BallerinaException("error in set string to statement: " + e.getMessage(), e);
-        }
-    }
-
-    public static void setDoubleValue(PreparedStatement stmt, BValue value, int index, int direction, int sqlType) {
-        Double val = null;
-        if (value != null) {
-            String strValue = value.stringValue();
-            if (!strValue.isEmpty()) {
-                try {
-                    val = Double.parseDouble(strValue);
-                } catch (NumberFormatException e) {
-                    throw new BallerinaException("invalid value for double: " + strValue);
-                }
-            }
-        }
-        try {
-            if (Constants.QueryParamDirection.IN == direction) {
-                if (val == null) {
-                    stmt.setNull(index + 1, sqlType);
-                } else {
-                    stmt.setDouble(index + 1, val);
-                }
-            } else if (Constants.QueryParamDirection.INOUT == direction) {
-                if (val == null) {
-                    stmt.setNull(index + 1, sqlType);
-                } else {
-                    stmt.setDouble(index + 1, val);
-                }
-                ((CallableStatement) stmt).registerOutParameter(index + 1, sqlType);
-            } else if (Constants.QueryParamDirection.OUT == direction) {
-                ((CallableStatement) stmt).registerOutParameter(index + 1, sqlType);
-            } else {
-                throw new BallerinaException("invalid direction for the parameter with index: " + index);
-            }
-        } catch (SQLException e) {
-            throw new BallerinaException("error in set double to statement: " + e.getMessage(), e);
-        }
-    }
-
-    public static void setNumericValue(PreparedStatement stmt, BValue value, int index, int direction, int sqlType) {
-        BigDecimal val = null;
-        if (value != null) {
-            String strValue = value.stringValue();
-            if (!strValue.isEmpty()) {
-                try {
-                    val = new BigDecimal(strValue);
-                } catch (NumberFormatException e) {
-                    throw new BallerinaException("invalid value for numeric: " + strValue);
-                }
-            }
-        }
-        try {
-            if (Constants.QueryParamDirection.IN == direction) {
-                if (val == null) {
-                    stmt.setNull(index + 1, sqlType);
-                } else {
-                    stmt.setBigDecimal(index + 1, val);
-                }
-            } else if (Constants.QueryParamDirection.INOUT == direction) {
-                if (val == null) {
-                    stmt.setNull(index + 1, sqlType);
-                } else {
-                    stmt.setBigDecimal(index + 1, val);
-                }
-                ((CallableStatement) stmt).registerOutParameter(index + 1, sqlType);
-            } else if (Constants.QueryParamDirection.OUT == direction) {
-                ((CallableStatement) stmt).registerOutParameter(index + 1, sqlType);
-            } else {
-                throw new BallerinaException("invalid direction for the parameter with index: " + index);
-            }
-        } catch (SQLException e) {
-            throw new BallerinaException("error in set numeric value to statement: " + e.getMessage(), e);
-        }
-    }
-
-    public static void setBooleanValue(PreparedStatement stmt, BValue value, int index, int direction, int sqlType) {
-        Boolean val = null;
-        if (value != null) {
-            String strValue = value.stringValue();
-            if (!strValue.isEmpty()) {
-                val = Boolean.valueOf(strValue);
-            }
-        }
-        try {
-            if (Constants.QueryParamDirection.IN == direction) {
-                if (val == null) {
-                    stmt.setNull(index + 1, sqlType);
-                } else {
-                    stmt.setBoolean(index + 1, val);
-                }
-            } else if (Constants.QueryParamDirection.INOUT == direction) {
-                if (val == null) {
-                    stmt.setNull(index + 1, sqlType);
-                } else {
-                    stmt.setBoolean(index + 1, val);
-                }
-                ((CallableStatement) stmt).registerOutParameter(index + 1, sqlType);
-            } else if (Constants.QueryParamDirection.OUT == direction) {
-                ((CallableStatement) stmt).registerOutParameter(index + 1, sqlType);
-            } else {
-                throw new BallerinaException("invalid direction for the parameter with index: " + index);
-            }
-        } catch (SQLException e) {
-            throw new BallerinaException("error in set boolean value to statement: " + e.getMessage(), e);
-        }
-    }
-
-    public static void setTinyIntValue(PreparedStatement stmt, BValue value, int index, int direction, int sqlType) {
-        Byte val = null;
-        if (value != null) {
-            String strValue = value.stringValue();
-            if (!strValue.isEmpty()) {
-                try {
-                    val = Byte.valueOf(strValue);
-                } catch (NumberFormatException e) {
-                    throw new BallerinaException("invalid value for byte: " + strValue);
-                }
-            }
-        }
-        try {
-            if (Constants.QueryParamDirection.IN == direction) {
-                if (val == null) {
-                    stmt.setNull(index + 1, sqlType);
-                } else {
-                    stmt.setByte(index + 1, val);
-                }
-            } else if (Constants.QueryParamDirection.INOUT == direction) {
-                if (val == null) {
-                    stmt.setNull(index + 1, sqlType);
-                } else {
-                    stmt.setByte(index + 1, val);
-                }
-                ((CallableStatement) stmt).registerOutParameter(index + 1, sqlType);
-            } else if (Constants.QueryParamDirection.OUT == direction) {
-                ((CallableStatement) stmt).registerOutParameter(index + 1, sqlType);
-            } else {
-                throw new BallerinaException("invalid direction for the parameter with index: " + index);
-            }
-        } catch (SQLException e) {
-            throw new BallerinaException("error in set tinyint value to statement: " + e.getMessage(), e);
-        }
-    }
-
-    public static void setBigIntValue(PreparedStatement stmt, BValue value, int index, int direction, int sqlType) {
-        Long val = null;
-        if (value != null) {
-            String strValue = value.stringValue();
-            if (!strValue.isEmpty()) {
-                try {
-                    val = Long.parseLong(strValue);
-                } catch (NumberFormatException e) {
-                    throw new BallerinaException("invalid value for bigint: " + strValue);
-                }
-            }
-        }
-        try {
-            if (Constants.QueryParamDirection.IN == direction) {
-                if (val == null) {
-                    stmt.setNull(index + 1, sqlType);
-                } else {
-                    stmt.setLong(index + 1, val);
-                }
-            } else if (Constants.QueryParamDirection.INOUT == direction) {
-                if (val == null) {
-                    stmt.setNull(index + 1, sqlType);
-                } else {
-                    stmt.setLong(index + 1, val);
-                }
-                ((CallableStatement) stmt).registerOutParameter(index + 1, sqlType);
-            } else if (Constants.QueryParamDirection.OUT == direction) {
-                ((CallableStatement) stmt).registerOutParameter(index + 1, sqlType);
-            } else {
-                throw new BallerinaException("invalid direction for the parameter with index: " + index);
-            }
-        } catch (SQLException e) {
-            throw new BallerinaException("error in set bigint value to statement: " + e.getMessage(), e);
-        }
-    }
-
-    public static void setRealValue(PreparedStatement stmt, BValue value, int index, int direction, int sqlType) {
-        Float val = null;
-        if (value != null) {
-            String strValue = value.stringValue();
-            if (!strValue.isEmpty()) {
-                try {
-                    val = Float.parseFloat(strValue);
-                } catch (NumberFormatException e) {
-                    throw new BallerinaException("invalid value for float: " + strValue);
-                }
-            }
-        }
-        try {
-            if (Constants.QueryParamDirection.IN == direction) {
-                if (val == null) {
-                    stmt.setNull(index + 1, sqlType);
-                } else {
-                    stmt.setFloat(index + 1, val);
-                }
-            } else if (Constants.QueryParamDirection.INOUT == direction) {
-                if (val == null) {
-                    stmt.setNull(index + 1, sqlType);
-                } else {
-                    stmt.setFloat(index + 1, val);
-                }
-                ((CallableStatement) stmt).registerOutParameter(index + 1, sqlType);
-            } else if (Constants.QueryParamDirection.OUT == direction) {
-                ((CallableStatement) stmt).registerOutParameter(index + 1, sqlType);
-            } else {
-                throw new BallerinaException("invalid direction for the parameter, index: " + index);
-            }
-        } catch (SQLException e) {
-            throw new BallerinaException("error in set float value to statement." + e.getMessage(), e);
-        }
-    }
-
-    public static void setDateValue(PreparedStatement stmt, BValue value, int index, int direction, int sqlType) {
-        Date val = null;
-        if (value != null) {
-            if (value instanceof BMap && value.getType().getName().equals(Constants.STRUCT_TIME) && value.getType()
-                    .getPackagePath().equals(Constants.STRUCT_TIME_PACKAGE)) {
-                BValue timeVal = ((BMap<String, BValue>) value).get(Constants.STRUCT_TIME_FIELD);
-                long time = ((BInteger) timeVal).intValue();
-                val = new Date(time);
-            } else if (value instanceof BInteger) {
-                val = new Date(((BInteger) value).intValue());
-            } else if (value instanceof BString) {
-                val = SQLDatasourceUtils.convertToDate(value.stringValue());
-            } else {
-                throw new BallerinaException("invalid input type for date parameter with index: " + index);
-            }
-        }
-        try {
-            if (Constants.QueryParamDirection.IN == direction) {
-                if (val == null) {
-                    stmt.setNull(index + 1, sqlType);
-                } else {
-                    stmt.setDate(index + 1, val);
-                }
-            } else if (Constants.QueryParamDirection.INOUT == direction) {
-                if (val == null) {
-                    stmt.setNull(index + 1, sqlType);
-                } else {
-                    stmt.setDate(index + 1, val);
-                }
-                ((CallableStatement) stmt).registerOutParameter(index + 1, sqlType);
-            } else if (Constants.QueryParamDirection.OUT == direction) {
-                ((CallableStatement) stmt).registerOutParameter(index + 1, sqlType);
-            } else {
-                throw new BallerinaException("invalid direction for the parameter with index: " + index);
-            }
-        } catch (SQLException e) {
-            throw new BallerinaException("error in set date value to statement: " + e.getMessage(), e);
-        }
-    }
-
-    public static void setTimeStampValue(PreparedStatement stmt, BValue value, int index, int direction, int sqlType,
-            Calendar utcCalendar) {
-        Timestamp val = null;
-        if (value != null) {
-            if (value instanceof BMap && value.getType().getName().equals(Constants.STRUCT_TIME) && value.getType()
-                    .getPackagePath().equals(Constants.STRUCT_TIME_PACKAGE)) {
-                BValue timeVal = ((BMap<String, BValue>) value).get(Constants.STRUCT_TIME_FIELD);
-                long time = ((BInteger) timeVal).intValue();
-                val = new Timestamp(time);
-            } else if (value instanceof BInteger) {
-                val = new Timestamp(((BInteger) value).intValue());
-            } else if (value instanceof BString) {
-                val = SQLDatasourceUtils.convertToTimeStamp(value.stringValue());
-            } else {
-                throw new BallerinaException("invalid input type for timestamp parameter with index: " + index);
-            }
-        }
-        try {
-            if (Constants.QueryParamDirection.IN == direction) {
-                if (val == null) {
-                    stmt.setNull(index + 1, sqlType);
-                } else {
-                    stmt.setTimestamp(index + 1, val, utcCalendar);
-                }
-            } else if (Constants.QueryParamDirection.INOUT == direction) {
-                if (val == null) {
-                    stmt.setNull(index + 1, sqlType);
-                } else {
-                    stmt.setTimestamp(index + 1, val, utcCalendar);
-                }
-                ((CallableStatement) stmt).registerOutParameter(index + 1, sqlType);
-            } else if (Constants.QueryParamDirection.OUT == direction) {
-                ((CallableStatement) stmt).registerOutParameter(index + 1, sqlType);
-            } else {
-                throw new BallerinaException("invalid direction for the parameter, index: " + index);
-            }
-        } catch (SQLException e) {
-            throw new BallerinaException("error in set timestamp value to statement: " + e.getMessage(), e);
-        }
-    }
-
-    public static void setTimeValue(PreparedStatement stmt, BValue value, int index, int direction, int sqlType,
-            Calendar utcCalendar) {
-        Time val = null;
-        if (value != null) {
-            if (value instanceof BMap && value.getType().getName().equals(Constants.STRUCT_TIME) && value.getType()
-                    .getPackagePath().equals(Constants.STRUCT_TIME_PACKAGE)) {
-                BValue timeVal = ((BMap<String, BValue>) value).get(Constants.STRUCT_TIME_FIELD);
-                long time = ((BInteger) timeVal).intValue();
-                val = new Time(time);
-            } else if (value instanceof BInteger) {
-                val = new Time(((BInteger) value).intValue());
-            } else if (value instanceof BString) {
-                val = SQLDatasourceUtils.convertToTime(value.stringValue());
-            }
-        }
-        try {
-            if (Constants.QueryParamDirection.IN == direction) {
-                if (val == null) {
-                    stmt.setNull(index + 1, sqlType);
-                } else {
-                    stmt.setTime(index + 1, val, utcCalendar);
-                }
-            } else if (Constants.QueryParamDirection.INOUT == direction) {
-                if (val == null) {
-                    stmt.setNull(index + 1, sqlType);
-                } else {
-                    stmt.setTime(index + 1, val, utcCalendar);
-                }
-                ((CallableStatement) stmt).registerOutParameter(index + 1, sqlType);
-            } else if (Constants.QueryParamDirection.OUT == direction) {
-                ((CallableStatement) stmt).registerOutParameter(index + 1, sqlType);
-            } else {
-                throw new BallerinaException("invalid direction for the parameter with index: " + index);
-            }
-        } catch (SQLException e) {
-            throw new BallerinaException("error in set timestamp value to statement: " + e.getMessage(), e);
-        }
-    }
-
-    public static void setBinaryValue(PreparedStatement stmt, BValue value, int index, int direction, int sqlType) {
-        byte[] val = getByteArray(value);
-        try {
-            if (Constants.QueryParamDirection.IN == direction) {
-                if (val == null) {
-                    stmt.setNull(index + 1, sqlType);
-                } else {
-                    stmt.setBinaryStream(index + 1, new ByteArrayInputStream(val), val.length);
-                }
-            } else if (Constants.QueryParamDirection.INOUT == direction) {
-                if (val == null) {
-                    stmt.setNull(index + 1, sqlType);
-                } else {
-                    stmt.setBinaryStream(index + 1, new ByteArrayInputStream(val), val.length);
-                }
-                ((CallableStatement) stmt).registerOutParameter(index + 1, sqlType);
-            } else if (Constants.QueryParamDirection.OUT == direction) {
-                ((CallableStatement) stmt).registerOutParameter(index + 1, sqlType);
-            } else {
-                throw new BallerinaException("invalid direction for the parameter with index: " + index);
-            }
-        } catch (SQLException e) {
-            throw new BallerinaException("error in set binary value to statement: " + e.getMessage(), e);
-        }
-    }
-
-    public static void setBlobValue(PreparedStatement stmt, BValue value, int index, int direction, int sqlType) {
-        byte[] val = getByteArray(value);
-        try {
-            if (Constants.QueryParamDirection.IN == direction) {
-                if (val == null) {
-                    stmt.setNull(index + 1, sqlType);
-                } else {
-                    stmt.setBlob(index + 1, new ByteArrayInputStream(val), val.length);
-                }
-            } else if (Constants.QueryParamDirection.INOUT == direction) {
-                if (val == null) {
-                    stmt.setNull(index + 1, sqlType);
-                } else {
-                    stmt.setBlob(index + 1, new ByteArrayInputStream(val), val.length);
-                }
-                ((CallableStatement) stmt).registerOutParameter(index + 1, sqlType);
-            } else if (Constants.QueryParamDirection.OUT == direction) {
-                ((CallableStatement) stmt).registerOutParameter(index + 1, sqlType);
-            } else {
-                throw new BallerinaException("invalid direction for the parameter with index: " + index);
-            }
-        } catch (SQLException e) {
-            throw new BallerinaException("error in set binary value to statement: " + e.getMessage(), e);
-        }
-    }
-
-    private static Integer obtainIntegerValue(BValue value) {
-        if (value != null) {
-            String strValue = value.stringValue();
-            if (!strValue.isEmpty()) {
-                try {
-                    return Integer.parseInt(strValue);
-                } catch (NumberFormatException e) {
-                    throw new BallerinaException("invalid value for integer: " + strValue);
-                }
-            }
-        }
-        return null;
-    }
-
-    private static byte[] getByteArray(BValue value) {
-        byte[] val = null;
-        if (value instanceof BValueArray) {
-            val = ((BValueArray) value).getBytes();
-            val = Arrays.copyOfRange(val, 0, (int) value.size());
-        } else if (value instanceof BString) {
-            val = getBytesFromBase64String(value.stringValue());
-        }
-        return val;
-    }
-
-    public static void setClobValue(PreparedStatement stmt, BValue value, int index, int direction, int sqlType) {
-        BufferedReader val = null;
-        if (value != null) {
-            val = new BufferedReader(new StringReader(value.stringValue()));
-        }
-        try {
-            if (Constants.QueryParamDirection.IN == direction) {
-                if (val == null) {
-                    stmt.setNull(index + 1, sqlType);
-                } else {
-                    stmt.setClob(index + 1, val, value.stringValue().length());
-                }
-            } else if (Constants.QueryParamDirection.INOUT == direction) {
-                if (val == null) {
-                    stmt.setNull(index + 1, sqlType);
-                } else {
-                    stmt.setClob(index + 1, val, value.stringValue().length());
-                }
-                ((CallableStatement) stmt).registerOutParameter(index + 1, sqlType);
-            } else if (Constants.QueryParamDirection.OUT == direction) {
-                ((CallableStatement) stmt).registerOutParameter(index + 1, sqlType);
-            } else {
-                throw new BallerinaException("invalid direction for the parameter with index: " + index);
-            }
-        } catch (SQLException e) {
-            throw new BallerinaException("error in set binary value to statement: " + e.getMessage(), e);
-        }
-    }
-
-    public static void setNClobValue(PreparedStatement stmt, BValue value, int index, int direction, int sqlType) {
-        BufferedReader val = null;
-        if (value != null) {
-            val = new BufferedReader(new StringReader(value.stringValue()));
-        }
-        try {
-            if (Constants.QueryParamDirection.IN == direction) {
-                if (val == null) {
-                    stmt.setNull(index + 1, sqlType);
-                } else {
-                    stmt.setNClob(index + 1, val, value.stringValue().length());
-                }
-            } else if (Constants.QueryParamDirection.INOUT == direction) {
-                if (val == null) {
-                    stmt.setNull(index + 1, sqlType);
-                } else {
-                    stmt.setNClob(index + 1, val, value.stringValue().length());
-                }
-                ((CallableStatement) stmt).registerOutParameter(index + 1, sqlType);
-            } else if (Constants.QueryParamDirection.OUT == direction) {
-                ((CallableStatement) stmt).registerOutParameter(index + 1, sqlType);
-            } else {
-                throw new BallerinaException("invalid direction for the parameter with index: " + index);
-            }
-        } catch (SQLException e) {
-            throw new BallerinaException("error in set binary value to statement: " + e.getMessage(), e);
-        }
-    }
-
-    public static void setRefCursorValue(PreparedStatement stmt, int index, int direction, String databaseProductName) {
-        try {
-            if (Constants.QueryParamDirection.OUT == direction) {
-                if (Constants.DatabaseNames.ORACLE.equals(databaseProductName)) {
-                    // Since oracle does not support general java.sql.Types.REF_CURSOR in manipulating ref cursors it
-                    // is required to use oracle.jdbc.OracleTypes.CURSOR here. In order to avoid oracle driver being
-                    // a runtime dependency always, we have directly used the value(-10) of general oracle.jdbc
-                    // .OracleTypes.CURSOR here.
-                    ((CallableStatement) stmt).registerOutParameter(index + 1, ORACLE_CURSOR_TYPE);
-                } else {
-                    ((CallableStatement) stmt).registerOutParameter(index + 1, Types.REF_CURSOR);
-                }
-            } else {
-                throw new BallerinaException("invalid direction for the parameter with index: " + index);
-            }
-        } catch (SQLException e) {
-            throw new BallerinaException("error in setting ref cursor value to statement: " + e.getMessage(), e);
-        }
-    }
-
-    public static void setArrayValue(Connection conn, PreparedStatement stmt, BValue value, int index, int direction,
-            int sqlType, String databaseProductName) {
-        Object[] arrayData = getArrayData(value);
-        Object[] arrayValue = (Object[]) arrayData[0];
-        String structuredSQLType = (String) arrayData[1];
-        try {
-            if (Constants.QueryParamDirection.IN == direction) {
-                setArrayValue(arrayValue, conn, stmt, index, sqlType, databaseProductName, structuredSQLType);
-            } else if (Constants.QueryParamDirection.INOUT == direction) {
-                setArrayValue(arrayValue, conn, stmt, index, sqlType, databaseProductName, structuredSQLType);
-                registerArrayOutParameter(stmt, index, sqlType, structuredSQLType, databaseProductName);
-            } else if (Constants.QueryParamDirection.OUT == direction) {
-                registerArrayOutParameter(stmt, index, sqlType, structuredSQLType, databaseProductName);
-            } else {
-                throw new BallerinaException("invalid direction for the parameter with index: " + index);
-            }
-        } catch (SQLException e) {
-            throw new BallerinaException("error in set array value to statement: " + e.getMessage(), e);
-        }
-    }
-
-    public static void setNullObject(PreparedStatement stmt, int index) {
-        try {
-            stmt.setObject(index + 1, null);
-        } catch (SQLException e) {
-            throw new BallerinaException("error in set null to parameter with index: " + index);
-        }
-    }
-
-    private static Object[] getArrayData(BValue value) {
-        if (value == null || value.getType().getTag() != TypeTags.ARRAY_TAG) {
-            return new Object[] { null, null };
-        }
-        int typeTag = ((BArrayType) value.getType()).getElementType().getTag();
-        Object[] arrayData;
-        int arrayLength;
-        switch (typeTag) {
-        case TypeTags.INT_TAG:
-            arrayLength = (int) ((BValueArray) value).size();
-            arrayData = new Long[arrayLength];
-            for (int i = 0; i < arrayLength; i++) {
-                arrayData[i] = ((BValueArray) value).getInt(i);
-            }
-            return new Object[] { arrayData, Constants.SQLDataTypes.BIGINT };
-        case TypeTags.FLOAT_TAG:
-            arrayLength = (int) ((BValueArray) value).size();
-            arrayData = new Double[arrayLength];
-            for (int i = 0; i < arrayLength; i++) {
-                arrayData[i] = ((BValueArray) value).getFloat(i);
-            }
-            return new Object[] { arrayData, Constants.SQLDataTypes.DOUBLE };
-        case TypeTags.DECIMAL_TAG:
-            arrayLength = (int) ((BValueArray) value).size();
-            arrayData = new BigDecimal[arrayLength];
-            for (int i = 0; i < arrayLength; i++) {
-                arrayData[i] = ((BValueArray) value).getRefValue(i).value();
-            }
-            return new Object[] { arrayData, Constants.SQLDataTypes.DECIMAL };
-        case TypeTags.STRING_TAG:
-            arrayLength = (int) ((BValueArray) value).size();
-            arrayData = new String[arrayLength];
-            for (int i = 0; i < arrayLength; i++) {
-                arrayData[i] = ((BValueArray) value).getString(i);
-            }
-            return new Object[] { arrayData, Constants.SQLDataTypes.VARCHAR };
-        case TypeTags.BOOLEAN_TAG:
-            arrayLength = (int) ((BValueArray) value).size();
-            arrayData = new Boolean[arrayLength];
-            for (int i = 0; i < arrayLength; i++) {
-                arrayData[i] = ((BValueArray) value).getBoolean(i) > 0;
-            }
-            return new Object[] { arrayData, Constants.SQLDataTypes.BOOLEAN };
-        case TypeTags.ARRAY_TAG:
-            BType elementType = ((BArrayType) ((BArrayType) value.getType()).getElementType()).getElementType();
-            if (elementType.getTag() == TypeTags.BYTE_TAG) {
-                arrayLength = (int) ((BValueArray) value).size();
-                arrayData = new Blob[arrayLength];
-                for (int i = 0; i < arrayLength; i++) {
-                    arrayData[i] = ((BValueArray) value).getByte(i);
-                }
-                return new Object[] { arrayData, Constants.SQLDataTypes.BLOB };
-            } else {
-                throw new BallerinaException("unsupported data type for array parameter");
-            }
-        default:
-            throw new BallerinaException("unsupported data type for array parameter");
-        }
-    }
-
-    public static void setUserDefinedValue(Connection conn, PreparedStatement stmt, BValue value, int index,
-            int direction, int sqlType) {
-        try {
-            Object[] structData = getStructData(value, conn);
-            Object[] dataArray = (Object[]) structData[0];
-            String structuredSQLType = (String) structData[1];
-            if (Constants.QueryParamDirection.IN == direction) {
-                if (dataArray == null) {
-                    stmt.setNull(index + 1, sqlType);
-                } else {
-                    Struct struct = conn.createStruct(structuredSQLType, dataArray);
-                    stmt.setObject(index + 1, struct);
-                }
-            } else if (Constants.QueryParamDirection.INOUT == direction) {
-                if (dataArray == null) {
-                    stmt.setNull(index + 1, sqlType);
-                } else {
-                    Struct struct = conn.createStruct(structuredSQLType, dataArray);
-                    stmt.setObject(index + 1, struct);
-                }
-                ((CallableStatement) stmt).registerOutParameter(index + 1, sqlType, structuredSQLType);
-            } else if (Constants.QueryParamDirection.OUT == direction) {
-                ((CallableStatement) stmt).registerOutParameter(index + 1, sqlType, structuredSQLType);
-            } else {
-                throw new BallerinaException("invalid direction for the parameter with index: " + index);
-            }
-        } catch (SQLException e) {
-            throw new BallerinaException("error in set struct value to statement: " + e.getMessage(), e);
-        }
-    }
-
-    private static Object[] getStructData(BValue value, Connection conn) throws SQLException {
-        if (value == null || (value.getType().getTag() != TypeTags.OBJECT_TYPE_TAG
-                && value.getType().getTag() != TypeTags.RECORD_TYPE_TAG)) {
-            return new Object[] { null, null };
-        }
-        String structuredSQLType = value.getType().getName().toUpperCase(Locale.getDefault());
-        Map<String, BField> structFields = ((BStructureType) value.getType()).getFields();
-        int fieldCount = structFields.size();
-        Object[] structData = new Object[fieldCount];
-        Iterator<BField> fieldIterator = structFields.values().iterator();
-        for (int i = 0; i < fieldCount; ++i) {
-            BField field = fieldIterator.next();
-            BValue bValue = ((BMap<String, BValue>) value).get(field.fieldName);
-            int typeTag = field.getFieldType().getTag();
-            switch (typeTag) {
-            case TypeTags.INT_TAG:
-                structData[i] = ((BInteger) bValue).intValue();
-                break;
-            case TypeTags.FLOAT_TAG:
-                structData[i] = ((BFloat) bValue).floatValue();
-                break;
-            case TypeTags.STRING_TAG:
-                structData[i] = bValue.stringValue();
-                break;
-            case TypeTags.BOOLEAN_TAG:
-                structData[i] = ((BBoolean) bValue).booleanValue();
-                break;
-            case TypeTags.DECIMAL_TAG:
-                structData[i] = ((BDecimal) bValue).decimalValue();
-                break;
-            case TypeTags.ARRAY_TAG:
-                BType elementType = ((BArrayType) field.getFieldType()).getElementType();
-                if (elementType.getTag() == TypeTags.BYTE_TAG) {
-                    structData[i] = ((BValueArray) bValue).getBytes();
-                    break;
-                } else {
-                    throw new BallerinaException("unsupported data type for struct parameter: " + structuredSQLType);
-                }
-            case TypeTags.OBJECT_TYPE_TAG:
-            case TypeTags.RECORD_TYPE_TAG:
-                Object structValue = bValue;
-                if (structValue instanceof BMap) {
-                    Object[] internalStructData = getStructData((BMap<String, BValue>) structValue, conn);
-                    Object[] dataArray = (Object[]) internalStructData[0];
-                    String internalStructType = (String) internalStructData[1];
-                    structValue = conn.createStruct(internalStructType, dataArray);
-                }
-                structData[i] = structValue;
-                break;
-            default:
-                throw new BallerinaException("unsupported data type for struct parameter: " + structuredSQLType);
-            }
-        }
-        return new Object[] { structData, structuredSQLType };
+    public void checkAndObserveSQLError(Context context, String message) {
+        Optional<ObserverContext> observerContext = ObserveUtils.getObserverContextOfCurrentFrame(context);
+        observerContext.ifPresent(ctx -> {
+            ctx.addProperty(ObservabilityConstants.PROPERTY_ERROR, Boolean.TRUE);
+            ctx.addProperty(ObservabilityConstants.PROPERTY_ERROR_MESSAGE, message);
+        });
     }
 
     /**
@@ -981,32 +208,6 @@ public class SQLDatasourceUtils {
             return TypeKind.RECORD;
         default:
             return TypeKind.NONE;
-        }
-    }
-
-    public static String getSQLType(BType value) {
-        int tag = value.getTag();
-        switch (tag) {
-        case TypeTags.INT_TAG:
-            return Constants.SQLDataTypes.BIGINT;
-        case TypeTags.STRING_TAG:
-            return Constants.SQLDataTypes.VARCHAR;
-        case TypeTags.FLOAT_TAG:
-            return Constants.SQLDataTypes.DOUBLE;
-        case TypeTags.BOOLEAN_TAG:
-            return Constants.SQLDataTypes.BOOLEAN;
-        case TypeTags.DECIMAL_TAG:
-            return Constants.SQLDataTypes.DECIMAL;
-        case TypeTags.ARRAY_TAG:
-            if (((BArrayType) value).getElementType().getTag() == TypeTags.BYTE_TAG) {
-                return Constants.SQLDataTypes.BINARY;
-            } else {
-                throw new BallerinaException("Array data type as direct value is supported only " +
-                        "with byte type elements, use sql:Parameter " + value.getName());
-            }
-        default:
-            throw new BallerinaException(
-                    "unsupported data type as direct value for sql operation, use sql:Parameter: " + value.getName());
         }
     }
 
@@ -1134,10 +335,10 @@ public class SQLDatasourceUtils {
         return null;
     }
 
-    public static BError getSQLConnectorError(Context context, Throwable throwable) {
+    public static BError getSQLConnectorError(Context context, Throwable throwable, String messagePrefix) {
         String detailedErrorMessage =
             throwable.getMessage() != null ? throwable.getMessage() : Constants.DATABASE_ERROR_MESSAGE;
-        return getSQLConnectorError(context, detailedErrorMessage);
+        return getSQLConnectorError(context, messagePrefix + detailedErrorMessage);
     }
 
     public static BError getSQLConnectorError(Context context, String detailedErrorMessage) {
@@ -1341,30 +542,6 @@ public class SQLDatasourceUtils {
         return dbOptions.isEmpty() ? jdbcUrl.toString() : jdbcUrl.append(dbOptions).toString();
     }
 
-    private static void registerArrayOutParameter(PreparedStatement stmt, int index, int sqlType,
-            String structuredSQLType, String databaseProductName) throws SQLException {
-        if (databaseProductName.equals(Constants.DatabaseNames.POSTGRESQL)) {
-            ((CallableStatement) stmt).registerOutParameter(index + 1, sqlType);
-        } else {
-            ((CallableStatement) stmt).registerOutParameter(index + 1, sqlType, structuredSQLType);
-        }
-    }
-
-    private static void setArrayValue(Object[] arrayValue, Connection conn, PreparedStatement stmt, int index,
-            int sqlType, String databaseProductName, String structuredSQLType) throws SQLException {
-        if (arrayValue == null) {
-            stmt.setNull(index + 1, sqlType);
-        } else {
-            // In POSTGRESQL, need to pass "float8" to indicate DOUBLE value.
-            if (Constants.SQLDataTypes.DOUBLE.equals(structuredSQLType) && Constants.DatabaseNames.POSTGRESQL
-                    .equals(databaseProductName)) {
-                structuredSQLType = POSTGRES_DOUBLE;
-            }
-            Array array = conn.createArrayOf(structuredSQLType, arrayValue);
-            stmt.setArray(index + 1, array);
-        }
-    }
-
     private static BMap<String, BValue> createSQLClient(Context context,
             SQLDatasource.SQLDatasourceParams sqlDatasourceParams) {
         SQLDatasource sqlDatasource = sqlDatasourceParams.getPoolOptionsWrapper()
@@ -1399,14 +576,6 @@ public class SQLDatasourceUtils {
             throw new BallerinaException("invalid type for datetime data: " + type);
         }
         return datetimeString.toString();
-    }
-
-    private static byte[] getBytesFromBase64String(String base64Str) {
-        try {
-            return Base64.getDecoder().decode(base64Str.getBytes(Charset.defaultCharset()));
-        } catch (Exception e) {
-            throw new BallerinaException("error in processing base64 string: " + e.getMessage(), e);
-        }
     }
 
     private static byte[] getBase64Encode(String st) {
@@ -1479,203 +648,6 @@ public class SQLDatasourceUtils {
             dateString.append("0");
         }
         dateString.append(calendar.get(Calendar.DAY_OF_MONTH));
-    }
-
-    private static Date convertToDate(String source) {
-        // the lexical form of the date is '-'? yyyy '-' mm '-' dd zzzzzz?
-        if ((source == null) || source.trim().equals("")) {
-            return null;
-        }
-        source = source.trim();
-        Calendar calendar = Calendar.getInstance();
-        calendar.clear();
-        calendar.setLenient(false);
-        if (source.startsWith("-")) {
-            source = source.substring(1);
-            calendar.set(Calendar.ERA, GregorianCalendar.BC);
-        }
-        if (source.length() >= 10) {
-            if ((source.charAt(4) != '-') || (source.charAt(7) != '-')) {
-                throw new BallerinaException("invalid date format: " + source);
-            }
-            int year = Integer.parseInt(source.substring(0, 4));
-            int month = Integer.parseInt(source.substring(5, 7));
-            int day = Integer.parseInt(source.substring(8, 10));
-            int timeZoneOffSet = TimeZone.getDefault().getRawOffset();
-            if (source.length() > 10) {
-                String restpart = source.substring(10);
-                timeZoneOffSet = getTimeZoneOffset(restpart);
-                calendar.set(Calendar.DST_OFFSET, 0); //set the day light offset only if the time zone is present
-            }
-            calendar.set(Calendar.YEAR, year);
-            calendar.set(Calendar.MONTH, month - 1);
-            calendar.set(Calendar.DAY_OF_MONTH, day);
-            calendar.set(Calendar.ZONE_OFFSET, timeZoneOffSet);
-        } else {
-            throw new BallerinaException("invalid date string to parse: " + source);
-        }
-        return new Date(calendar.getTime().getTime());
-    }
-
-    private static Time convertToTime(String source) {
-        //lexical representation of the time is hh ':' mm ':' ss ('.' s+)? (zzzzzz)?
-        if ((source == null) || source.trim().equals("")) {
-            return null;
-        }
-        source = source.trim();
-        Calendar calendar = Calendar.getInstance();
-        calendar.clear();
-        calendar.setLenient(false);
-        if (source.length() >= 8) {
-            if ((source.charAt(2) != ':') || (source.charAt(5) != ':')) {
-                throw new BallerinaException("invalid time format: " + source);
-            }
-            int hour = Integer.parseInt(source.substring(0, 2));
-            int minite = Integer.parseInt(source.substring(3, 5));
-            int second = Integer.parseInt(source.substring(6, 8));
-            int miliSecond = 0;
-            int timeZoneOffSet = TimeZone.getDefault().getRawOffset();
-            if (source.length() > 8) {
-                String rest = source.substring(8);
-                int[] offsetData = getTimeZoneWithMilliSeconds(rest);
-                miliSecond = offsetData[0];
-                timeZoneOffSet = offsetData[1];
-                calendar.set(Calendar.DST_OFFSET, 0); //set the day light offset only if the time zone is present
-            }
-            calendar.set(Calendar.HOUR_OF_DAY, hour);
-            calendar.set(Calendar.MINUTE, minite);
-            calendar.set(Calendar.SECOND, second);
-            calendar.set(Calendar.MILLISECOND, miliSecond);
-            calendar.set(Calendar.ZONE_OFFSET, timeZoneOffSet);
-        } else {
-            throw new BallerinaException("time string can not be less than 8 characters: " + source);
-        }
-        return new Time(calendar.getTimeInMillis());
-    }
-
-    private static Timestamp convertToTimeStamp(String source) {
-        //lexical representation of the date time is '-'? yyyy '-' mm '-' dd 'T' hh ':' mm ':' ss ('.' s+)? (zzzzzz)?
-        if ((source == null) || source.trim().equals("")) {
-            return null;
-        }
-        source = source.trim();
-        Calendar calendar = Calendar.getInstance();
-        calendar.clear();
-        calendar.setLenient(false);
-        if (source.startsWith("-")) {
-            source = source.substring(1);
-            calendar.set(Calendar.ERA, GregorianCalendar.BC);
-        }
-        if (source.length() >= 19) {
-            if ((source.charAt(4) != '-') || (source.charAt(7) != '-') || (source.charAt(10) != 'T') || (
-                    source.charAt(13) != ':') || (source.charAt(16) != ':')) {
-                throw new BallerinaException("invalid datetime format: " + source);
-            }
-            int year = Integer.parseInt(source.substring(0, 4));
-            int month = Integer.parseInt(source.substring(5, 7));
-            int day = Integer.parseInt(source.substring(8, 10));
-            int hour = Integer.parseInt(source.substring(11, 13));
-            int minite = Integer.parseInt(source.substring(14, 16));
-            int second = Integer.parseInt(source.substring(17, 19));
-            long miliSecond = 0;
-            int timeZoneOffSet = TimeZone.getDefault().getRawOffset();
-            if (source.length() > 19) {
-                String rest = source.substring(19);
-                int[] offsetData = getTimeZoneWithMilliSeconds(rest);
-                miliSecond = offsetData[0];
-                int calculatedtimeZoneOffSet = offsetData[1];
-                if (calculatedtimeZoneOffSet != -1) {
-                    timeZoneOffSet = calculatedtimeZoneOffSet;
-                }
-                calendar.set(Calendar.DST_OFFSET, 0); //set the day light offset only if the time zone is present
-            }
-            calendar.set(Calendar.YEAR, year);
-            calendar.set(Calendar.MONTH, month - 1);
-            calendar.set(Calendar.DAY_OF_MONTH, day);
-            calendar.set(Calendar.HOUR_OF_DAY, hour);
-            calendar.set(Calendar.MINUTE, minite);
-            calendar.set(Calendar.SECOND, second);
-            calendar.set(Calendar.MILLISECOND, (int) miliSecond);
-            calendar.set(Calendar.ZONE_OFFSET, timeZoneOffSet);
-        } else {
-            throw new BallerinaException("datetime string can not be less than 19 characters: " + source);
-        }
-        return new Timestamp(calendar.getTimeInMillis());
-    }
-
-    private static int[] getTimeZoneWithMilliSeconds(String fractionStr) {
-        int miliSecond = 0;
-        int timeZoneOffSet = -1;
-        if (fractionStr.startsWith(".")) {
-            int milliSecondPartLength;
-            if (fractionStr.endsWith("Z")) { //timezone is given as Z
-                timeZoneOffSet = 0;
-                String fractionPart = fractionStr.substring(1, fractionStr.lastIndexOf("Z"));
-                miliSecond = Integer.parseInt(fractionPart);
-                milliSecondPartLength = fractionPart.trim().length();
-            } else {
-                int lastIndexOfPlus = fractionStr.lastIndexOf("+");
-                int lastIndexofMinus = fractionStr.lastIndexOf("-");
-                if ((lastIndexOfPlus > 0) || (lastIndexofMinus > 0)) { //timezone +/-hh:mm
-                    String timeOffSetStr;
-                    if (lastIndexOfPlus > 0) {
-                        timeOffSetStr = fractionStr.substring(lastIndexOfPlus + 1);
-                        String fractionPart = fractionStr.substring(1, lastIndexOfPlus);
-                        miliSecond = Integer.parseInt(fractionPart);
-                        milliSecondPartLength = fractionPart.trim().length();
-                        timeZoneOffSet = 1;
-                    } else {
-                        timeOffSetStr = fractionStr.substring(lastIndexofMinus + 1);
-                        String fractionPart = fractionStr.substring(1, lastIndexofMinus);
-                        miliSecond = Integer.parseInt(fractionPart);
-                        milliSecondPartLength = fractionPart.trim().length();
-                        timeZoneOffSet = -1;
-                    }
-                    if (timeOffSetStr.charAt(2) != ':') {
-                        throw new BallerinaException("invalid time zone format: " + fractionStr);
-                    }
-                    int hours = Integer.parseInt(timeOffSetStr.substring(0, 2));
-                    int minits = Integer.parseInt(timeOffSetStr.substring(3, 5));
-                    timeZoneOffSet = ((hours * 60) + minits) * 60000 * timeZoneOffSet;
-
-                } else { //no timezone
-                    miliSecond = Integer.parseInt(fractionStr.substring(1));
-                    milliSecondPartLength = fractionStr.substring(1).trim().length();
-                }
-            }
-            if (milliSecondPartLength != 3) {
-                // milisecond part represenst the fraction of the second so we have to
-                // find the fraction and multiply it by 1000. So if milisecond part
-                // has three digits nothing required
-                miliSecond = miliSecond * 1000;
-                for (int i = 0; i < milliSecondPartLength; i++) {
-                    miliSecond = miliSecond / 10;
-                }
-            }
-        } else {
-            timeZoneOffSet = getTimeZoneOffset(fractionStr);
-        }
-        return new int[] { miliSecond, timeZoneOffSet };
-    }
-
-    private static int getTimeZoneOffset(String timezoneStr) {
-        int timeZoneOffSet;
-        if (timezoneStr.startsWith("Z")) { //GMT timezone
-            timeZoneOffSet = 0;
-        } else if (timezoneStr.startsWith("+") || timezoneStr.startsWith("-")) { //timezone with offset
-            if (timezoneStr.charAt(3) != ':') {
-                throw new BallerinaException("invalid time zone format:" + timezoneStr);
-            }
-            int hours = Integer.parseInt(timezoneStr.substring(1, 3));
-            int minits = Integer.parseInt(timezoneStr.substring(4, 6));
-            timeZoneOffSet = ((hours * 60) + minits) * 60000;
-            if (timezoneStr.startsWith("-")) {
-                timeZoneOffSet = timeZoneOffSet * -1;
-            }
-        } else {
-            throw new BallerinaException("invalid prefix for timezone: " + timezoneStr);
-        }
-        return timeZoneOffSet;
     }
 
     public static List<ColumnDefinition> getColumnDefinitions(ResultSet rs) throws SQLException {
